feat(#138): fix all qulice suggestions
