+++ src/test/java/org/eolang/jeo/representation/directives/DirectivesMethodTest.java
@@ -38,8 +38,10 @@ import org.xembly.Xembler;
 /**
  * Test case for {@link org.eolang.jeo.representation.directives.DirectivesMethod}.
  * We create {@link org.eolang.jeo.representation.directives.DirectivesMethod} only in the context
- * of using {@link org.eolang.jeo.representation.directives.DirectivesClass} in other words, {@link org.eolang.jeo.representation.directives.DirectivesMethod} can't be createad
- * without {@link org.eolang.jeo.representation.directives.DirectivesClass} and it is the main reason why in all the test we create
+ * of using {@link org.eolang.jeo.representation.directives.DirectivesClass} in other words,
+ * {@link org.eolang.jeo.representation.directives.DirectivesMethod} can't be createad
+ * without {@link org.eolang.jeo.representation.directives.DirectivesClass} and it is the main
+ * reason why in all the test we create
  * {@link org.eolang.jeo.representation.directives.DirectivesClass}.
  *
  * @since 0.1.0
